[Improve]Ingress uses different versions of API according to different K8s versions.
