+++ streampark-flink/streampark-flink-kubernetes/src/test/scala/org/apache/streampark/flink/kubernetes/FlinkRestJsonTest.scala
@@ -17,6 +17,7 @@
 package org.apache.streampark.flink.kubernetes
 
 import org.apache.streampark.flink.kubernetes.helper.KubernetesDeploymentHelper
+import org.apache.streampark.flink.kubernetes.ingress.{IngressStrategyV1, IngressStrategyV1beta1}
 import org.apache.streampark.flink.kubernetes.watcher.{Checkpoint, FlinkRestJmConfigItem, FlinkRestOverview, JobDetails}
 
 import com.google.common.base.Charsets
@@ -294,7 +295,7 @@ class FlinkRestJsonTest {
         |]
         |""".stripMargin
 
-    val ingressMeta = IngressMeta.as(json)
+    val ingressMeta = new IngressStrategyV1().IngressMeta.as(json)
     println(ingressMeta.get)
   }
